State machine for sending hid reports every 8 ms, when not zero.
Based on the reverse engineerd SpaceNavigator this commit introduces a little state machine, which sends new data every 8 ms. This is not to be configured by every user and therefore not in the config_sample.h.
If no data is to sent, the zero data are sent three times before stopping the transmissions.
Changed the HID descriptor to a symmetric logical range of +/- 350. It has been -342 to +350, despite the comment saying it to be -350 all the time. 0xFEA2 = 350, not 0xFEAA.
Therefore I also needed the key state instead of the single key events.