Turning on LED by PC driver with new USB Interface which receives data. (#52)
This merge request implements a very basic LED.

The clue is, that the LED is turned by the PC driver, which can set the LED output!

To use the LED, copy the corresponding lines from the config_sample.h into your config.

The whole USB Interface was rewritten to have an extra endpoint, which is directed to the space mouse.

This is done by refactoring the hidInterface into a pluggableUSB class SpaceMouseHID. This is done based on the work presented by arandomdev in the arduino forum and by comparing the results to NicoHoods HID library, which is not necessary as a dependency in the project now, but is a great inspiration.

The LED routine has been found out by sniffing on a Space Navigator with wireshark. The raw wireshark data and some explanation is written here SpaceNavigator.md.
The "calibrate" button in the driver is unfortunately not working, as we are emulating a space mouse pro wireless (cabled). The driver issues the LED blinking sequence. But the final call to calibrate is not send for a space mouse pro, but only for the Space Navigator.